Simplify run_all_benchmarks to only run needed benchmarks
This changes the run to avoid running unnecessary benchmarks.

The single-threaded runs every time, so having a whole loop for it is
unnecessary.
The baseline (non-XSMM) only needs to be run once, since the XSMM
options make no difference. We set to scalar/block pointer semantics to
gather all data.

Options:
 * datatype: f32, bf16, bf8 (default f32)
 * threads: default cores per socket (no HT)
 * baseline: run xsmm benchmarks or not